Added tests for FileInstance

+++ classes/fileinstance.rb
@@ -12,17 +12,18 @@ module Email
         end
 
         def to_s
-            self.existance
-            file = File.open(@file_path, 'r')
-            file_text = file.read
-            file.close
-            file_text
+            if self.existance?
+                file = File.open(@file_path, 'r')
+                file_text = file.read
+                file.close
+                file_text
+            else
+                raise 'NoFileFound'
+            end
         end
 
-        def existance
-            unless File.exists?(file_path)
-                abort(file_path + "doesn't exist")
-            end
+        def existance?
+            File.exists?(file_path)
         end
 
         def write(file_contents)
@@ -32,13 +33,17 @@ module Email
         end
 
         def to_hash
-            if File.extname(@file_path) == '.json'
-                json_hash = JSON.parse(self.to_s)
-            elsif ['.yaml', '.yml'].include? File.extname(@file_path)
-                yaml_hash = YAML.load(@file_path.to_s)
-            else
-                abort(@file_path + 'must be a JSON or YAML file')
-            end
+            if self.existance?
+                if File.extname(@file_path) == '.json'
+                    json_hash = JSON.parse(self.to_s)
+                elsif ['.yaml', '.yml'].include? File.extname(@file_path)
+                    yaml_hash = YAML.load(self.to_s)
+                else
+                    raise 'Must be a JSON or YAML file'
+                end
+             else
+                 raise 'NoFileFound'
+             end
         end
 
         def _symbolize(obj)
